Cleaned up mgl::graphics
