+++ mgl/graphics/inc/mgl_graphics/commands/draw.hpp
@@ -4,6 +4,7 @@
 #include "mgl_graphics/command.hpp"
 
 #include "mgl_platform/api/buffers.hpp"
+#include "mgl_platform/api/render_api.hpp"
 
 namespace mgl::graphics
 {
@@ -24,7 +25,11 @@ public:
         , m_instance_count(instance_count)
     { }
 
-    void execute() override;
+    void execute() override final
+    {
+      mgl::platform::api::render_api::render_call(
+          m_vertex_buffer, m_index_buffer, m_count, m_offset, m_mode);
+    }
 
 private:
     mgl::platform::api::vertex_buffer_ref m_vertex_buffer;
@@ -42,7 +47,35 @@ public:
         : m_batch(batch)
     { }
 
-    void execute() override;
+    void execute() override final
+    {
+      MGL_CORE_ASSERT(false, "Not implemented");
+      // if(m_batch->index_buffer())
+      // {
+      // mgl::platform::api::vertex_array_ref vao =
+      //     mgl::platform::api::create_vertex_array(m_batch->vertex_buffer()->api(),
+      //                                             m_batch->vertex_buffer()->layout(),
+      //                                             m_batch->index_buffer()->api(),
+      //                                             m_batch->index_buffer()->element_size(),
+      //                                             m_batch->mode());
+      // vao->allocate();
+      // mgl::platform::api::draw_batch(vao, m_batch->get());
+      // vao->deallocate();
+      // }
+      // else
+      // {
+      // mgl::platform::api::vertex_array_ref vao =
+      //     mgl::platform::api::create_vertex_array(m_batch->vertex_buffer()->api(),
+      //                                             m_batch->vertex_buffer()->layout(),
+      //                                             nullptr,
+      //                                             4,
+      //                                             m_batch->mode());
+
+      // vao->allocate();
+      // mgl::platform::api::draw_batch(vao, m_batch->get());
+      // vao->deallocate();
+      // }
+    }
 
 private:
     batch_ref m_batch;
