+++ mgl/graphics/inc/mgl_graphics/commands/state.hpp
@@ -2,6 +2,8 @@
 
 #include "mgl_graphics/command.hpp"
 
+#include "mgl_platform/api/render_api.hpp"
+
 namespace mgl::graphics
 {
   class enable_state : public render_command
@@ -11,7 +13,7 @@ public:
         : m_flags(flags)
     { }
 
-    void execute() override;
+    void execute() override final { mgl::platform::api::render_api::enable_state(m_flags); }
 
 private:
     int m_flags;
@@ -24,7 +26,7 @@ public:
         : m_flags(flags)
     { }
 
-    void execute() override;
+    void execute() override final { mgl::platform::api::render_api::disable_state(m_flags); }
 
 private:
     int m_flags;
